14 work 170822

+++ controllers/cards.js
@@ -1,22 +1,25 @@
 const Card = require('../models/card');
 const { ValidationError } = require('../errors/ValidationError');
 const { NoValidId } = require('../errors/NoValidId');
+const { NoPermission } = require('../errors/NoPermission');
 const { CastError } = require('../errors/CastError');
 
-const returnCards = (req, res) => {
-  Card.find({})
+const returnCards = (req, res, next) => {
+  Card.find()
     .then((card) => res.send(card))
-    .catch(() => { res.status(500).send({ message: 'Произошла ошибка' }); });
+    .catch((err) => next(err));
 };
 
 const createCards = (req, res, next) => {
   const { name, link } = req.body;
 
   Card.create({ name, link, owner: req.user._id })
-    .then((card) => res.status(201).send(card))
+    .then((card) => {
+      res.status(201).send(card);
+    })
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        next(new ValidationError('400 - Переданы некорректные данные при создания карточки'));
+        next(new ValidationError('400 - Переданы некорректные данные в метод создания карточки'));
       } else {
         next(err);
       }
@@ -24,14 +27,31 @@ const createCards = (req, res, next) => {
 };
 
 const deleteCards = (req, res, next) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(new Error('NoValidId'))
-    .then((cardDeleted) => res.send(cardDeleted))
-    .catch((err) => {
-      if (err.message === 'NoValidId') {
+  Card.findById(req.params.cardId)
+    .then((card) => {
+      const owner = card.owner.toHexString();
+      if (!card) {
         next(new NoValidId('404 - Карточка с указанным _id не найдена'));
-      } else if (err.name === 'CastError') {
-        next(new CastError('400 —  Получение каточки с некорректным id'));
+      } else if (owner === req.user._id) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .orFail(new Error('NoValidId'))
+          .then((cardDeleted) => res.send(cardDeleted))
+          .catch((err) => {
+            if (err.message === 'NoValidId') {
+              next(new NoValidId('404 - Карточка с указанным _id не найдена'));
+            } else {
+              next(err);
+            }
+          });
+      } else {
+        next(new NoPermission('403 — попытка удалить чужую карточку;'));
+      }
+    })
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        next(new CastError('400 —  Карточка с указанным _id не найдена'));
+      } else if (err.name === 'TypeError') {
+        next(new NoValidId('404 - Удаление карточки с несуществующим в БД id'));
       } else {
         next(err);
       }
@@ -41,7 +61,7 @@ const deleteCards = (req, res, next) => {
 const setLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
-    { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
+    { $addToSet: { likes: req.user._id } },
     { new: true },
   )
     .orFail(new Error('NoValidId'))
@@ -50,11 +70,9 @@ const setLike = (req, res, next) => {
     })
     .catch((err) => {
       if (err.message === 'NoValidId') {
-        next(new NoValidId('404 — Передан несуществующий _id'));
-      } else if (err.name === 'CastError') {
-        next(new CastError('400 —  Передан некорректный id'));
+        next(new NoValidId('404 — Передан несуществующий _id карточки'));
       } else {
-        next(err);
+        next(new Error('Ошибка. Что-то пошло не так...'));
       }
     });
 };
@@ -62,7 +80,7 @@ const setLike = (req, res, next) => {
 const unsetLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
-    { $pull: { likes: req.user._id } }, // добавить _id в массив, если его там нет
+    { $pull: { likes: req.user._id } },
     { new: true },
   )
     .orFail(new Error('NoValidId'))
@@ -71,11 +89,9 @@ const unsetLike = (req, res, next) => {
     })
     .catch((err) => {
       if (err.message === 'NoValidId') {
-        next(new NoValidId('404 — Передан несуществующий _id'));
-      } else if (err.name === 'CastError') {
-        next(new CastError('400 —  Передан некорректный id'));
+        next(new NoValidId('404 — Передан несуществующий _id карточки'));
       } else {
-        next(err);
+        next(new Error('Ошибка. Что-то пошло не так...'));
       }
     });
 };
